This is a working version that will display people and their offspring. The problem that I had with the Query results as explained by Asseem:
Cypher query results are never just an array of nodes -- they're an
array of results, where a "result" is a dictionary of variable name to
the corresponding (node/rel/path/etc.) object.

So in your example, your Cypher query says RETURN m, so in your for
loop, you need to access results[i]['m'], not just results[i].

This fixed the problem in the persons.js model. I then added code inthe
view to display a list of offsping for a selected person/